Day2/100
Keep splitting the linked list into two parts from the middle using the split list function (a function which uses two pointers, one of them travels one node at a time and the other two nodes. At the end, the first node is at a middle point so we can split the list into two halves) and just keep merging the two parts of the list by seeing which of the parts data is smaller and keep on returning the merged list.